nginx chnage

+++ nginx.sh
@@ -10,8 +10,11 @@ sudo service nginx stop
 # Copy the Nginx configuration file to the appropriate directory
 sudo cp ./nginx.conf /etc/nginx/sites-available/evp
 
-# Create a symbolic link to enable the configuration file
-sudo ln -s /etc/nginx/sites-available/evp /etc/nginx/sites-enabled/
+# Check if the symbolic link already exists before creating one
+if [ ! -f /etc/nginx/sites-enabled/evp ]; then
+    sudo ln -s /etc/nginx/sites-available/evp /etc/nginx/sites-enabled/
+fi
 
 # Start Nginx
 sudo service nginx start
+
